% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_gitbook}
\alias{thesis_gitbook}
\title{Creates an R Markdown gitbook Thesis document}
\usage{
thesis_gitbook(...)
}
\arguments{
\item{...}{arguments to \code{\link[bookdown:gitbook]{bookdown::gitbook()}}}
}
\value{
A gitbook webpage
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a webpage version of the thesis.
}
\examples{
\dontrun{
 output: thesisdownUNIL::thesis_gitbook
}
}
\seealso{
\code{\link[thesisdown:thesis_gitbook]{thesisdown::thesis_gitbook()}}, \code{\link[bookdown:gitbook]{bookdown::gitbook()}}
}
